requirements.txt
discord.py==2.4.0
python-dotenv==0.10.1
